Memory Optimizations! - Disable shadow and remove unused directional light If we do enable shadows, it'd be better to limit it to a singular light source. Fewer render passes, better performance.
- Replaced Point Light with Directional Light (tried to match lighting
as best as possible)
Point Lights have to calculate the bounds the light reaches.
Directional is global (less calculating)

- Disabled Mesh colliders on Plateau -> Rocks -> FireCircle + RandomRocks
Mesh colliders take up too much memory if concave, should at least
use convex property, but need to separate rocks into separate meshes
first.

- Removed Scene transition script from helix in fireplace
Only needed in GameManager

- Reduced Camera clipping plane far from 4000 to 2500 for Fireplace.
From 4000 to 1000 in Spiral.
Will try pairing with a fog post-process to hide any possible
clipping

- Removed MKGlowLite, using built-in GlowShaders instead.
- Disabled Menu Handler
- Load in Background set for Audio files and streaming for main
tracks reducing memory load
Load In Background: If enabled, the audio clip will be loading
in the background without causing stalls on the main thread.
This is off by default in order to ensure the standard Unity
behavior where all AudioClips have finished loading when the
scene starts playing.
Streaming: Decode sounds on the fly. This method uses a minimal
amount of memory to buffer compressed data that is incrementally
read from the disk and decoded on the fly.

- Removed alpha source for non-transparent textures. Should
ideally replace textures with RGB instead of RGBA profile type
to save on memory.

- Manually set the Max Size on textures! Compresses textures
so they're only as big as they're needed.

- Non Power of 2 textures set to nearest instead of largest
- Disabled MipMaps on larger textures
Generate Mip Maps: Select this to enable mip-map generation.
Mip maps are smaller versions of the texture that get used
when the texture is very small on screen.

- Transcode 360 video to 4096 x 2048 dimensions and ensured
textures have same dimensions